Moved FullCourseSelectionController
