+++ app/views/candidate_interface/course_choices/full_course_selection/new.html.erb
